refactor(guard): F617 — fix cross-domain imports to use infra/types.js contract (MSA baseline=0)
